// src/models/grid.model.ts
export declare type SortOrder = "descend" | "ascend" | null;
export declare type textcasetype = "upper" | "lower" | "sentence";
export declare type PagintionPosition =
  | "topRight"
  // | "topCenter"
  // | "topLeft"
  // | "bottomCenter"
  | "bottomRight";
// | "bottomLeft";

//RecordType is the type of data being passed which is generic
export declare type RowClassName<RecordType> = (
  record: RecordType,
  index: number,
  indent: number
) => string;

export interface ColumnFilterItem {
  text: React.ReactNode;
  value: string | number | boolean;
}
// ...
export interface FilterDropdownProps {
  prefixCls: string;
  setSelectedKeys: (selectedKeys: React.Key[]) => void;
  selectedKeys: React.Key[];
  confirm: () => void;
  clearFilters?: () => void;
  filters?: ColumnFilterItem[];
  visible: boolean;
}
export interface Column<RecordType> {
  displayTitle: string;
  headerCellSelection?: boolean;
  textCase?: textcasetype;
  className?: string;
  showToolTip?: boolean;
  formatter?: JSX.Element;
  showDecimals?: boolean;
  cellWrapper?: (props: any) => JSX.Element;
  toolTip?: (props) => JSX.Element;
  customContent?: (props) => JSX.Element;
  isFilterable?: boolean;
  key: string;
  fixed?: "left" | "right";
  dataType?: "string" | "number" | "date";
  enableIntellisense?: boolean;
  complexObject?: string; //eg if lob is an object and value is name property => complexObject:"name" else null/undefined
  render?: (record: RecordType) => JSX.Element;
  width?: number;
  minWidth?: number;
  pixelWidth?: number;
  position: number;
  hidden: boolean;
  title?: React.ReactNode; //HTML for th
  isClickable?: boolean;

  //DEPRECATED
  onCellClick?: (record, key) => void;

  componentToOpenOnClickingCell?: (props) => JSX.Element;
  isFilterble?: boolean;

  sorter?: {
    multiple?: number;
  };
  sortDirections?: SortOrder[];
  sortOrder?: SortOrder;
  defaultSortOrder?: SortOrder;
  showSorterTooltip?: boolean;
  filters?: ColumnFilterItem[];
  filteredValue?: any;
  filterDropdown?:
    | React.ReactNode
    | ((props: FilterDropdownProps) => React.ReactNode);
  filterMultiple?: boolean;

  filterIcon?: (filtered: any) => JSX.Element;
  // onFilter?: (value: string | number | boolean, record: RecordType) => boolean;

  filterDropdownVisible?: boolean;
  onFilterDropdownVisibleChange?: (visible: boolean) => void;
}